init: wait for docker for bridge IP
It can happen that we're running before the docker service is even up
yet. In those cases, the docker0 bridge might not be created and up yet.
Let's wait for the docker service to finish initializing before trying
to get the docker0 IP.

Also, we now switch from using /sys/ and `ip` to using nmcli, which is
probably safer and easier to parse.