stay in the grub2 menu across updates
We no longer want the boot menu entry to disappear after an explicit or
implicit `grub2-mkconfig`. To do this, we change the mechanism by which
the boot menu gets added:
1. The `bootentry` script adds/removes hooks inside the `/etc/grub.d/`
   directory. The hooks are positioned before and after the ostree hook.
   The `bootentry` script is now what the kickstart has to call (or what
   users must call if upgrading manually) in order to have the bootmenu
   added.
2. The `grub2_hook` script is run once before the ostree hook to find
   the latest BLS entry and create a new devmode entry based on that. It
   is also run once after the ostree hook to purge out the devmode
   entry. Otherwise on the next bootswap, ostree will read it in and
   create a new renamed entry in the future loader dir, so we wouldn't
   be able to recognize it and purge it.

This way, everytime the user does an upgrade/rebase/rollback, the
devmode entry is updated so that it uses the same tree as the new
default entry.