grub2_hook: skip entries with no atomic-devmode
If the user is rebasing, rolling back, or downgrading into a tree
without atomic-devmode, we don't want to create a devmode entry which
uses that tree. Otherwise, users booting it will experience cloud-init
hanging.

We tweak the algorithm so that the highest version *with* atomic-devmode
is chosen as the template.