Introduce 'showpasswd' alias
It is possible for a user of devmode to scroll the top pane of the
devmode session so that the root password disappears from the console.
If the user had not recorded the root password or not changed it or
otherwise was incapable of remembering the password, they are left in a
bind.

This change introduces an alias, `showpasswd`, that will print the root
password to the screen again.  Since the root password is automatically
generated and stored in `/run/atomic-devmode-root`, the alias just does
a `cat` on the file.

I also updated `init` to include a message about the `showpasswd`
command.