add cloud-init cleanup service
We want to make sure devmode doesn't cause future boots to act
differently. The issue is that metadata from an attached drive can
override our instance-id, thus causing per-instance modules to not run
on second boot. Let's just go to the nuclear option and wipe everything
clean as soon as cloud-init is done so that no semaphores are kept.

Note this is also documented in upstream cloud-init in
var-lib-cloud.txt:

> to clear out the current instance's data as if to force a "new run" on
> reboot do:
>   ( cd /var/lib/cloud/instance && sudo rm -Rf * )

This is essentially what our little service does.