init: error out if no IP address is acquired
Currently, if the system is unable to acquire an IP address, the
`get_external_ip` function will return `N\A`.  This is a successful
return and the check at line 69 will succeed and proceed to try and
download the Cockpit image.

This changes the behavior of `get_external_ip` to return `1` if it is
determined there is no IP address assigned to the system. This should
cause the init script to error and stop.

Fixes #3