bootcmd: expire root password
Cockpit 0.118 adds support for handling expired passwords. Let's make
use of this here so that users are forced to change the root password on
the first login into the Cockpit console.

This functionality requires the SSH daemon to allow
ChallengeResponseAuthentication. So we make sure that it's on for the
devmode boot.